# Up to 2x speedup on GPUs using memory efficient attention