--- a/docs/source/optimization/fp16.mdx
+++ b/docs/source/optimization/fp16.mdx
@@ -295,7 +295,8 @@ with torch.inference_mode():
 
 ## Memory Efficient Attention
 Recent work on optimizing the bandwitdh in the attention block have generated huge speed ups and gains in GPU memory usage. The most recent being Flash Attention (from @tridao, [code](https://github.com/HazyResearch/flash-attention), [paper](https://arxiv.org/pdf/2205.14135.pdf)) .
-Here are the speedups we obtain on a few Nvidia GPUs: 
+Here are the speedups we obtain on a few Nvidia GPUs when running the inference at 512x512 with a batch size of 1 (one prompt):
+ 
 | GPU              	| Base Attention FP16 	| Memory Efficient Attention FP16 	|
 |------------------	|---------------------	|---------------------------------	|
 | NVIDIA Tesla T4  	| 3.5it/s             	| 5.5it/s                         	|
@@ -325,6 +326,6 @@ pipe.enable_xformers_memory_efficient_attention()
 with torch.inference_mode(), torch.autocast("cuda", dtype=torch.float16):
     sample = pipe("a small cat")
 
-# optional: You can disable it via 
+# optional: You can disable it via
 # pipe.disable_xformers_memory_efficient_attention()
 ```