… src/diffusers/models/unet_2d_condition.py
… src/diffusers/pipelines/stable_diffusion/pipeline_stable_diffusion.py
… src/diffusers/pipelines/stable_diffusion/pipeline_stable_diffusion_img2img.py
… src/diffusers/pipelines/stable_diffusion/pipeline_stable_diffusion_inpaint.py
… src/diffusers/utils/import_utils.py