# Why ? 
While stable diffusion democratized the access to text to image generative models, it can still be relatively long to generate an image on consumer GPUs. The GPU memory requirements are also hindering the use of diffusion on small GPUs. 


# How ?
Recent work on optimizing the bandwitdh in the attention block have generated huge speed ups and gains in GPU memory usage. The most recent being Flash Attention (from @tridao, [code](https://github.com/HazyResearch/flash-attention), [paper](https://arxiv.org/pdf/2205.14135.pdf)) . 

In this PR we use the MemoryEfficientAttention implementation from [xformers](https://github.com/facebookresearch/xformers) (cc. @fmassa, @danthe3rd, @blefaudeux) to both speedup the cross-attention speed and decrease its GPU memory requirements. 

The memory efficient attention can be activated by setting the environment variable `USE_MEMORY_EFFICIENT_ATTENTION=1`
and installing the `xformers` library: 

```
pip install git+https://github.com/facebookresearch/xformers@51dd119#egg=xformers
```

This installation is a known pain point, there are two ways to improve that: 
  - xformers ships wheels
  - this dependency could be made optional in this repository

Thank you @tridao, @fmassa, @danthe3rd for the work on Flash Attention and its integration in xformers. 
Would it be possible to add a more optimised kernel for head-dim=40 which is the parameter used in stable diffusion. @blefaudeux and I would be happy to contribute :) 

Note: 
 - Masked cross-attention is not supported yet with this implementation but could be added in future work. 
 - This PR does not solve the xformers dependency just yet, any help appreciated

### Speedups on various GPUs with a 512x512 shape and running FP16: 
| GPU              	| Base Attention FP16 	| Memory Efficient Attention FP16 	|
|------------------	|---------------------	|---------------------------------	|
| NVIDIA Tesla T4  	| 3.5it/s             	| 5.5it/s                         	|
| NVIDIA 3060 RTX  	| 4.6it/s             	| 7.8it/s                         	|
| NVIDIA A10G      	| 8.88it/s            	| 15.6it/s                        	|
| NVIDIA RTX A6000 	| 11.7it/s            	| 21.09it/s                       	|
| A100-SXM-80GB    	| 18.7it/s            	| 29.5it/s                        	|
 
 
 ### How to test: 
I use the following setup: 
```
sudo docker run -it --gpus=all --ipc=host -v /home:/home nvcr.io/nvidia/pytorch:22.08-py3 bash

# Then 
pip install git+https://github.com/facebookresearch/xformers@51dd119#egg=xformers
pip install transformers ftfy scipy

# Followed by
cd PATH_TO_DIFFUSER_FOLDER
git checkout memory_efficient_attention
pip install -e . 
``` 
 
 Then create a python file, mine is named `test.py`, with the following code:
 ```
import torch
from diffusers import StableDiffusionPipeline


pipe = StableDiffusionPipeline.from_pretrained(
    "CompVis/stable-diffusion-v1-4", revision="fp16", torch_dtype=torch.float16, use_auth_token=True
).to("cuda")

with torch.inference_mode(), torch.autocast("cuda"):
    image = pipe("a small cat")
 ```
 
 Then run in the aforementioned docker container
 ```
 # Test without Memory Efficient Attention: 
 python test.py
 
 # Test with Memory Efficient Attention: 
 USE_MEMORY_EFFICIENT_ATTENTION=1 python test.py
 ```