@fmassa 2022-09-16
Hi @MatthieuTPHR ,

Nice PR!

On xFormers side, we are working on improving the packaging so that it can be more easily installed by users, while shipping the pre-compiled binaries as well.

We are also continuing to optimize the kernel for some configurations, we will keep the `K=40` in mind for the future.