@patrickvonplaten 2022-09-16
Hey @MatthieuTPHR,

Thanks a lot for opening the PR - it looks very cool! Trying it out now :-) 

Generally we're quite careful with not adding new dependencies to `diffusers`, but I think we might be able to make it a soft-dependency if the speed-up is big enough!