@danthe3rd 2022-09-16
Hi @MatthieuTPHR - this looks like a great improvement!

> Would it be possible to add a more optimised kernel for head-dim=40 which is the parameter used in stable diffusion. @blefaudeux and I would be happy to contribute :)

We've been improving the forward (including fairly recently, in https://github.com/facebookresearch/xformers/pull/388 for instance). Do you mind sharing the other parameters you use (datatype, sequence length, number of heads) - so we can add them to our benchmarks?