@blefaudeux 2022-09-16
> Hi @MatthieuTPHR - this looks like a great improvement!
> 
> > Would it be possible to add a more optimised kernel for head-dim=40 which is the parameter used in stable diffusion. @blefaudeux and I would be happy to contribute :)
> 
> We've been improving the forward (including fairly recently, in [facebookresearch/xformers#388](https://github.com/facebookresearch/xformers/pull/388) for instance). Do you mind sharing the other parameters you use (datatype, sequence length, number of heads) - so we can add them to our benchmarks?

typically with stable diffusion 512x512 yields a 64x64 latent space -> 4096 tokens (but higher res would be even better). 
After that the embedding is 320 over 8 heads, which yields a head dim of 40. For inference batch is typically 2 (conditioned and unconditioned diffusion), but can depend on the methods. So [2, 4096, 40] is a good baseline, [2, 16384, 40] is a bonus (1024 rendering)