@blefaudeux 2022-09-16
> I've tried running the code in this PR, but I'm getting the following error:
> 
> ```
> AttributeError: module 'triton.language' has no attribute 'constexpr'
> ```
> 
> when installing `xformers` with `pip install git+https://github.com/facebookresearch/xformers@51dd119#egg=xformers `
> 
> Do I need a specific version of triton cc @MatthieuTPHR ?

specifically `triton == 2.0.0.dev20220701` will work, it needs to be updated on xformers but that's WIP (the newer ones break a couple of kernels)