@patrickvonplaten 2022-09-16
I'm testing on a `NVIDIA TITAN RTX` on this branch with the following package dependencies:

```
- CUDA Version: 11.6
- torch: 1.12.1+cu102
- xformers: 0.0.13.dev (installed with `pip install git+https://github.com/facebookresearch/xformers@51dd119#egg=xformers`)
- triton: 2.0.0 (installed with `pip install triton==2.0.0.dev20220701`)
```

**Note**:

When I import xformers I'm getting:
```
libtorch_cuda_cu.so: cannot open shared object file: No such file or directory
WARNING:root:WARNING: libtorch_cuda_cu.so: cannot open shared object file: No such file or directory
Need to compile C++ extensions to get sparse attention suport. Please run python setup.py build develop
0.0.13.dev
```

With this setup I'm running: 

```python
from diffusers import StableDiffusionPipeline
import numpy as np
import torch

model_id = "CompVis/stable-diffusion-v1-4"
pipe = StableDiffusionPipeline.from_pretrained(
    model_id,
    use_auth_token=True,
)
pipe.to("cuda")

prompt = "A fantasy landscape, trending on artstation"
generator = torch.Generator(device="cuda").manual_seed(0)

with torch.autocast("cuda"):
    output = pipe(prompt=prompt, guidance_scale=7.5, generator=generator, output_type="np")

print(np.sum(np.abs(output.images[:3, :3, :3, :3])))
mem_bytes = torch.cuda.max_memory_allocated()
print(mem_bytes)
```

1. Without `xformers` simply on the current `main` branch. Without `xformers` I'm getting: 9.15iterations/sec
2. I'm running on **this** branch with:

```
export USE_MEMORY_EFFICIENT_ATTENTION=1
```

And I'm getting the exact same speed.

Any ideas what could be the problem here?