@fmassa 2022-09-16
Hi @patrickvonplaten 

It might look like the CUDA extensions are not being compiled when installing xformers.

can you try doing
```
import torch, xformers.ops

print(torch.ops.xformers.efficient_attention_forward_generic)
```
and see if ti prints something like `<OpOverloadPacket(op='xformers.efficient_attention_forward_generic')>` ?

If that doesn't print what I mentioned, there are a few options why this isn't being compiled:
A few questions:
- do you have CUDA setup in the machine you installed xformers?
- do you have a `nvcc` in the machine you used to install xformers?