@MatthieuTPHR 2022-09-16
@patrickvonplaten here is my full setup: 

Here is my full setup: 

```
sudo docker run -it --gpus=all --ipc=host -v /home:/home nvcr.io/nvidia/pytorch:22.08 bash

# Then 
pip install git+https://github.com/facebookresearch/xformers@51dd119#egg=xformers
pip install transformers ftfy scipy

# Followed by
cd PATH_TO_DIFFUSER_FOLDER
git checkout memory_efficient_attention
pip install -e . 
```

Once this is done I can notice the speedup.