@tridao 2022-09-16
Thank you @MatthieuTPHR, super exited to see ideas on fast & memory-efficient attention having an impact!