@patrickvonplaten 2022-09-16
Hey @fmassa,

When running:

```python
import torch, xformers.ops

print(torch.ops.xformers.efficient_attention_forward_generic)
```

I'm getting:

```
'_OpNamespace' object has no attribute 'efficient_attention_forward_generic'
```