@fmassa 2022-09-16
@patrickvonplaten this means that indeed xformers was compiled without the CUDA extensions.

> It would be really nice if we could somehow show the community that it's easy to install and use :-)

Yes, I totally agree, and we are working on that :-)

If you are not compiling xformers on a machine with CUDA (i.e., if you `pip install` in a machine without GPUs and then ssh to a machine with CPUs), you could try using `FORCE_CUDA=1 pip install git+https://github.com/facebookresearch/xformers@51dd119#egg=xformers` to try and compile xformers. Compiling it with its CUDA extensions should take a while, so it might be useful to pass `--verbose` to `pip` to see what are the things that are being done.