@TheLastBen 2022-09-18
Tested on GTX 1070ti :
Without Memory efficient cross attention at 512x512 : **1.78 it/s**

With Memory efficient cross attention at 512x512 : **2.34 it/s**

**+31%** increase in speed 

with optimized SD + Dogettx : at 512x512 : **2.0 it/s** (+12%  increase in speed)

With Only Doggettx modification, the speed isn't affected : **2.34 it/s**

Great work !